feat(admin-page): Add confirm to the delete buttons

+++ public/js/admin/admin.js
@@ -7,6 +7,9 @@ buttonsDelete.forEach(button => {
 	button.addEventListener('click', async e => {
 		const id = button.dataset.id ?? ''
 
+		const confirmDelete = confirm(`En verdad quieres borrar el item del id '${id}' ?`)
+		if (!confirmDelete) return
+
 		const response = await fetch(`/admin/delete/${id}`, {
 			method: 'DELETE',
 		})
